+++ blocklets/ai-studio/api/src/libs/env.ts
@@ -1,5 +1,6 @@
 import path from 'path';
 
+import { ServiceMode, getServiceModePermissionMap } from '@blocklet/ai-runtime/common';
 import config from '@blocklet/sdk/lib/config';
 import Joi from 'joi';
 
@@ -28,6 +29,14 @@ export const Config = {
 
   get usageReportThrottleTime() {
     return 30e3;
+  },
+
+  get serviceModePermissionMap() {
+    const { serviceMode } = config.env.preferences as {
+      serviceMode: ServiceMode;
+    };
+
+    return getServiceModePermissionMap(serviceMode);
   },
 };