+++ blocklets/ai-studio/src/components/file-editor/prompt-file/setting.tsx
@@ -1,4 +1,5 @@
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
+import { getSupportedModels } from '@blocklet/ai-runtime/common';
 import { PromptAssistantYjs } from '@blocklet/ai-runtime/types';
 import { ExpandMoreRounded, InfoOutlined } from '@mui/icons-material';
 import { Box, Collapse, FormLabel, Stack, Tooltip, Typography } from '@mui/material';
@@ -6,7 +7,6 @@ import isNil from 'lodash/isNil';
 import { useMemo, useState } from 'react';
 import { useAsync } from 'react-use';
 
-import { getSupportedModels } from '../../../../api/src/libs/common';
 import { useAssistantCompare, useProjectState } from '../../../pages/project/state';
 import WithAwareness from '../../awareness/with-awareness';
 import ModelSelectField, { brandIcon } from '../../selector/model-select-field';