+++ blocklets/ai-studio/src/locales/en.ts
@@ -575,6 +575,7 @@ export default flat({
   shareTip: 'Access link',
   refreshTip: 'Reembedding',
   loginAsAdminButton: 'Login as Admin to view projects',
+  login: 'Login to view projects',
   datasetDesc: 'Search for the data required by the context from the user dataset',
   types: 'Type',
   atLeastOne: 'Choose at least one type.',